ref(pretty): remove `--pretty` flag
- remove --pretty flag
- move types to utils
- update commands
